backend/api/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python
ultralytics
Pillow